chat

+++ src/components/route/PrivateRoute.tsx
@@ -10,22 +10,35 @@ import {
   setFriends,
   setUser,
 } from "@/redux/features/auth/auth.slice";
+import { useUpdateReadChatMessageMutation } from "@/redux/features/chat/chat.service";
+import {
+  addChatMessageSelectedHead,
+  addUserContactNew,
+  updateChatMessageSelected,
+  updateUserContactsByMessage,
+} from "@/redux/features/chat/chat.slice";
 import { setClientStomp } from "@/redux/features/global.slice";
 import { useLazyGetFriendsQuery } from "@/redux/features/user/user.service";
 import { useAppDispatch, useAppSelector } from "@/redux/hooks";
+import { Message } from "@/types/chat.type";
 import { API_URL, NotificationFriendRequest } from "@/types/contants.type";
 import { PostNotification } from "@/types/post.type";
 import { FriendRequestResponse } from "@/types/response.type";
 import User, { UserProfile } from "@/types/user.type";
+import { detectDeviceType, getIPAddress } from "@/utils/get-info.utils";
 import { getLocalStorageItem } from "@/utils/localstorage.util";
 import { getSessionData, setSessionData } from "@/utils/session.util";
-import { Client } from "@stomp/stompjs";
+import { Client, StompSubscription } from "@stomp/stompjs";
 import { useRouter } from "next/navigation";
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import { ToastContainer } from "react-toastify";
 
 const PrivateRoute = ({ children }: { children: React.ReactNode }) => {
   const user = useAppSelector((selector) => selector.auth.user);
+  const { userContactSelected, openChat, userContacts } = useAppSelector(
+    (selector) => selector.chat
+  );
+
   const accessToken = useAppSelector((selector) => selector.auth.accessToken);
   const { isDarkTheme } = useAppSelector((selector) => selector.global);
   const dispatch = useAppDispatch();
@@ -33,33 +46,54 @@ const PrivateRoute = ({ children }: { children: React.ReactNode }) => {
   const router = useRouter();
   const [relogin] = useLazyReloginQuery();
   const [getFriends] = useLazyGetFriendsQuery();
-
+  const [updateReadChatMessage] = useUpdateReadChatMessageMutation();
+  const clientStomp = useAppSelector((selector) => selector.global.clientStomp);
+  const subscriptions = useRef<StompSubscription[]>([]);
 
   const connectStomp = useCallback(() => {
-    console.log("accessToken", accessToken);
-    if (accessToken) {
+    if (subscriptions.current.length > 0) {
+      subscriptions.current.forEach((subscription) =>
+        subscription.unsubscribe()
+      );
+      subscriptions.current = [];
+    }
+
+    let ip_address: string | null = "";
+    const device_information: string = detectDeviceType();
+
+    getIPAddress().then((response) => {
+      ip_address = response;
+    });
+
+    if (accessToken && user?.id) {
       const header = {
         Authorization: `Bearer ${accessToken}`,
+        ip_address,
+        device_information,
       };
 
       const client = new Client({
         brokerURL: `ws://${API_URL._DOMAIN}/ws`,
-        debug: function (str) {
-          console.log(str);
-        },
+        // debug: function (str) {
+        //   console.log(str);
+        // },
       });
       client.onConnect = () => {
         if (user?.id) {
-          client.subscribe(`/user/${user.id}/topic/friendship`, (message) => {
-            const data: FriendRequestResponse = JSON.parse(message.body);
-            if (data.status === NotificationFriendRequest.CANCEL) {
-              dispatch(removeFriendRequest(data));
-            } else {
-              dispatch(addFriendRequestHead(data));
+          const subscribeFriendship = client.subscribe(
+            `/user/${user.id}/topic/friendship`,
+            (message) => {
+              const data: FriendRequestResponse = JSON.parse(message.body);
+              if (data.status === NotificationFriendRequest.CANCEL) {
+                dispatch(removeFriendRequest(data));
+              } else {
+                dispatch(addFriendRequestHead(data));
+              }
             }
-          });
+          );
+          subscriptions.current.push(subscribeFriendship);
 
-          client.subscribe(
+          const subscribePostNotification = client.subscribe(
             `/user/${user.id}/topic/post-notification`,
             (message) => {
               const data: PostNotification = JSON.parse(message.body);
@@ -70,6 +104,52 @@ const PrivateRoute = ({ children }: { children: React.ReactNode }) => {
               }
             }
           );
+          subscriptions.current.push(subscribePostNotification);
+
+          const subscribeMessage = client.subscribe(
+            `/user/${user.id}/topic/message`,
+            async (data) => {
+              const message: Message = JSON.parse(data.body);
+              dispatch(
+                addUserContactNew({
+                  message: message,
+                  userId: user.id,
+                })
+              );
+
+              if (
+                openChat &&
+                userContactSelected &&
+                message.sender.id === userContactSelected.id
+              ) {
+                dispatch(addChatMessageSelectedHead(message));
+                const response = await updateReadChatMessage(
+                  message.id
+                ).unwrap();
+                if (response.code === 201) {
+                  dispatch(updateChatMessageSelected(response.data));
+                  dispatch(updateUserContactsByMessage(response.data));
+                }
+              }
+            }
+          );
+          subscriptions.current.push(subscribeMessage);
+
+          const subscribeUpdateRead = client.subscribe(
+            `/user/${user.id}/topic/message/update-read`,
+            (data) => {
+              const message: Message = JSON.parse(data.body);
+              if (
+                openChat &&
+                userContactSelected &&
+                message.reciever.id === userContactSelected.id
+              ) {
+                dispatch(updateChatMessageSelected(message));
+              }
+              dispatch(updateUserContactsByMessage(message));
+            }
+          );
+          subscriptions.current.push(subscribeUpdateRead);
         }
       };
 
@@ -86,7 +166,14 @@ const PrivateRoute = ({ children }: { children: React.ReactNode }) => {
       client.activate();
       dispatch(setClientStomp(client));
     }
-  }, [dispatch, user, accessToken]);
+  }, [
+    dispatch,
+    user,
+    accessToken,
+    userContactSelected,
+    openChat,
+    updateReadChatMessage,
+  ]);
 
   useEffect(() => {
     const handleRelogin = async () => {
